+++ src/sentry/grouping/grouptype.py
@@ -3,18 +3,7 @@ from dataclasses import dataclass
 from sentry.issues.grouptype import GroupCategory, GroupType
 from sentry.models.group import DEFAULT_TYPE_ID
 from sentry.types.group import PriorityLevel
-from sentry.workflow_engine.handlers.detector.base import DetectorHandler
-from sentry.workflow_engine.models.data_source import DataPacket
-from sentry.workflow_engine.types import DetectorEvaluationResult, DetectorGroupKey, DetectorType
-
-
-class ErrorDetectorHandler(DetectorHandler[object]):
-    """Placeholder handler for error group types."""
-
-    def evaluate(
-        self, data_packet: DataPacket[object]
-    ) -> dict[DetectorGroupKey, DetectorEvaluationResult]:
-        return {}
+from sentry.workflow_engine.types import DetectorType
 
 
 @dataclass(frozen=True)
